probes: Adding ceph OSD
This commit is about adding Ceph OSDs as nodes.

On the agent side, it does extract the ceph osd metadata in the JSON and
encode it into Software.Ceph.OSD.metadata. This metadata is not supposed
to be graphed by the node but by the analyzer.

This behavior is needed as a single agent is able to describe the whole
cluster. So the analyzer will decode this metadata and add OSDs to
hosts.

OSDs nodes report links to theirs binded interfaces so until all
interfaces are present, the OSD graphing is deferred.

Once everything is properly graphed, the status of the metadata is kept
in memory to avoid regraphing an already graphed cluster. This feature
will save ressources by avoiding regraphing nodes.

Signed-off-by: Erwan Velu <erwan@redhat.com>